+++ lib/legion/cli/chat/tool_registry.rb
@@ -23,6 +23,7 @@ begin
   require 'legion/cli/chat/tools/knowledge_maintenance'
   require 'legion/cli/chat/tools/knowledge_stats'
   require 'legion/cli/chat/tools/summarize_traces'
+  require 'legion/cli/chat/tools/list_extensions'
 rescue LoadError => e
   Legion::Logging.debug("ToolRegistry ruby_llm not available, chat tools will not be registered: #{e.message}") if defined?(Legion::Logging)
 end
@@ -52,7 +53,8 @@ module Legion
                             Tools::RelateKnowledge,
                             Tools::KnowledgeMaintenance,
                             Tools::KnowledgeStats,
-                            Tools::SummarizeTraces
+                            Tools::SummarizeTraces,
+                            Tools::ListExtensions
                           ].freeze
                         else
                           [].freeze